Move Game Simulator subcomponent files to respective directories
