+++ src/pages/academy/gameSimulator/subcomponents/assetViewer/GameSimulatorAssetFileUploader.tsx
