+++ src/pages/academy/gameSimulator/subcomponents/assetViewer/GameSimulatorAssetSelection.tsx
