+++ src/pages/academy/gameSimulator/subcomponents/assetViewer/GameSimulatorAssetSelectionHelper.tsx
