+++ src/pages/academy/gameSimulator/subcomponents/assetViewer/GameSimulatorAssetViewer.tsx
