+++ src/pages/academy/gameSimulator/subcomponents/chapterPublisher/GameSimulatorChapterEditor.tsx
