+++ src/pages/academy/gameSimulator/subcomponents/chapterPublisher/GameSimulatorChapterSim.tsx
