+++ src/pages/academy/gameSimulator/subcomponents/chapterSimulator/GameSimulatorCheckpointSim.tsx
