+++ src/pages/academy/gameSimulator/subcomponents/chapterSimulator/GameSimulatorCheckpointTxtLoader.tsx
